feat(ida-88): pre action validation added
